Load zettlr main code
